+++ frontend/pages/index.tsx
@@ -1,21 +1,22 @@
 import React from "react";
-import { HeaderComponent } from "@/components/header-component";
-import { FooterComponent } from "@/components/footer-component";
 import { useWallet } from "@/contexts/WalletContext";
-import { KyosoComponent } from "@/components/kyoso-component";
+import { Footer } from "@/components/footer";
+import { Kyoso } from "@/components/kyoso";
+import Header from "@/components/header";
 
 export default function Home() {
   const { account, signer, connectWallet, disConnectWallet } = useWallet();
 
   return (
     <main className="flex min-h-screen flex-col items-center justify-between">
-      <KyosoComponent
+      <Header
         account={account}
         signer={signer}
         connectWallet={connectWallet}
         disConnectWallet={disConnectWallet}
       />
-      <FooterComponent />
+      <Kyoso />
+      <Footer />
     </main>
   );
 }
